claude-tools-chatbot/src/resources/initializerLambda/requirements.txt
boto3==1.34.117
botocore==1.34.117
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
